Update LHDR.rsl.txt
Modified the code to run based on a curve lookup table rather than a series of if statements. This optimization leads to slightly faster (~100ms) processing times. The change was suggested by Mike, the developer of GR2AE. Also cleaned up some of the documentation to better reflect what is actually occurring and some misunderstandings on how it executes.